requirements.txt
discord
pyrebase
python-dotenv
humanize